pyproject.toml
[build-system]
build-backend = "poetry.masonry.api"
requires = ["poetry>=0.12"]

[tool]
[tool.poetry]
authors = ["Your Name <you@example.com>"]
description = ""
name = "django-template"
version = "0.1.0"
[tool.poetry.dependencies]
Django = "^3.0"
python = "^3.8"
django-taggit = "^3.1.0"
django-taggit-labels = "^0.7.0"
django-search-views = "^0.3.7"
django-hitcount = "^1.3.5"
spkcspider = "^0.30"
[tool.poetry.dev-dependencies]
